compilation fixed for win32
svn path=/trunk/KDE/kdelibs/; revision=469371

+++ win/include/mingw/unistd.h
@@ -94,7 +94,7 @@ KDEWIN32_EXPORT pid_t setsid(void);
 #ifndef ICE_t
 #undef gethostname
 #define gethostname kde_gethostname
-KDEWIN32_EXPORT int kde_gethostname(char *__name, int __len);
+KDEWIN32_EXPORT int kde_gethostname(char *__name, size_t __len);
 #endif 
 
 KDEWIN32_EXPORT unsigned alarm(unsigned __secs );
